+++ contracts/governance/GovernorOHMegaDelegate.sol
@@ -10,30 +10,30 @@ contract GovernorOHMegaDelegate is GovernorOHMegaDelegateStorageV1, GovernorOHMe
 
     /// @notice The minimum setable proposal threshold
     /// @notice change from original contract
-    uint public constant MIN_PROPOSAL_THRESHOLDPERCENT = 5000; // 0.50% of sOHM supply : In ten-thosandaths 5000 = 0.50%
+    uint256 public constant MIN_PROPOSAL_THRESHOLDPERCENT = 5000; // 0.50% of sOHM supply : In ten-thosandaths 5000 = 0.50%
 
     /// @notice The maximum setable proposal threshold
     /// @notice change from original contract
-    uint public constant MAX_PROPOSAL_THRESHOLDPERCENT = 10000; // 1.00% of sOHM circulating supply : In ten-thosandaths 10000 = 1.00%
+    uint256 public constant MAX_PROPOSAL_THRESHOLDPERCENT = 10000; // 1.00% of sOHM circulating supply : In ten-thosandaths 10000 = 1.00%
 
     /// @notice The minimum setable voting period
-    uint public constant MIN_VOTING_PERIOD = 5760; // About 24 hours
+    uint256 public constant MIN_VOTING_PERIOD = 5760; // About 24 hours
 
     /// @notice The max setable voting period
-    uint public constant MAX_VOTING_PERIOD = 80640; // About 2 weeks
+    uint256 public constant MAX_VOTING_PERIOD = 80640; // About 2 weeks
 
     /// @notice The min setable voting delay
-    uint public constant MIN_VOTING_DELAY = 1;
+    uint256 public constant MIN_VOTING_DELAY = 1;
 
     /// @notice The max setable voting delay
-    uint public constant MAX_VOTING_DELAY = 40320; // About 1 week
+    uint256 public constant MAX_VOTING_DELAY = 40320; // About 1 week
 
     /// @notice The percent of sOHM in support of a proposal required in order for a quorum to be reached and for a vote to succeed
     /// @notice change from original contract
-    uint public constant quorumPercent = 40000; // In ten-thosandaths 40000 = 4.00%
+    uint256 public constant quorumPercent = 40000; // In ten-thosandaths 40000 = 4.00%
 
     /// @notice The maximum number of actions that can be included in a proposal
-    uint public constant proposalMaxOperations = 10; // 10 actions
+    uint256 public constant proposalMaxOperations = 10; // 10 actions
 
     /// @notice The EIP-712 typehash for the contract's domain
     bytes32 public constant DOMAIN_TYPEHASH = keccak256("EIP712Domain(string name,uint256 chainId,address verifyingContract)");
@@ -50,7 +50,7 @@ contract GovernorOHMegaDelegate is GovernorOHMegaDelegateStorageV1, GovernorOHMe
       * @param votingDelay_ The initial voting delay
       * @param proposalThreshold_ The initial proposal threshold percent in ten-thousandths
       */
-    function initialize(address timelock_, address sOHM_, address gOHM_, uint votingPeriod_, uint votingDelay_, uint proposalThreshold_) public {
+    function initialize(address timelock_, address sOHM_, address gOHM_, uint256 votingPeriod_, uint256 votingDelay_, uint256 proposalThreshold_) public {
         require(address(timelock) == address(0), "GovernorOHMega::initialize: can only initialize once");
         require(msg.sender == admin, " GovernorOHMega::initialize: admin only");
         require(timelock_ != address(0), "GovernorOHMega::initialize: invalid timelock address");
@@ -78,7 +78,7 @@ contract GovernorOHMegaDelegate is GovernorOHMegaDelegateStorageV1, GovernorOHMe
       * @param description String description of the proposal
       * @return Proposal id of new proposal
       */
-    function propose(address[] memory targets, uint[] memory values, string[] memory signatures, bytes[] memory calldatas, string memory description) public returns (uint) {
+    function propose(address[] memory targets, uint256[] memory values, string[] memory signatures, bytes[] memory calldatas, string memory description) public returns (uint256) {
         // Reject proposals before initiating as Governor
         require(initialProposalId != 0, "GovernorOHMega::propose: Governor OHMega not active");
         /// @notice change from original contract
@@ -87,15 +87,15 @@ contract GovernorOHMegaDelegate is GovernorOHMegaDelegateStorageV1, GovernorOHMe
         require(targets.length != 0, "GovernorOHMega::propose: must provide actions");
         require(targets.length <= proposalMaxOperations, "GovernorOHMega::propose: too many actions");
 
-        uint latestProposalId = latestProposalIds[msg.sender];
+        uint256 latestProposalId = latestProposalIds[msg.sender];
         if (latestProposalId != 0) {
           ProposalState proposersLatestProposalState = state(latestProposalId);
           require(proposersLatestProposalState != ProposalState.Active, "GovernorOHMega::propose: one live proposal per proposer, found an already active proposal");
           require(proposersLatestProposalState != ProposalState.Pending, "GovernorOHMega::propose: one live proposal per proposer, found an already pending proposal");
         }
 
-        uint startBlock = add256(block.number, votingDelay);
-        uint endBlock = add256(startBlock, votingPeriod);
+        uint256 startBlock = add256(block.number, votingDelay);
+        uint256 endBlock = add256(startBlock, votingPeriod);
 
         proposalCount++;
         Proposal memory newProposal = Proposal({
@@ -130,18 +130,18 @@ contract GovernorOHMegaDelegate is GovernorOHMegaDelegateStorageV1, GovernorOHMe
       * @notice Queues a proposal of state succeeded
       * @param proposalId The id of the proposal to queue
       */
-    function queue(uint proposalId) external {
+    function queue(uint256 proposalId) external {
         require(state(proposalId) == ProposalState.Succeeded, "GovernorOHMega::queue: proposal can only be queued if it is succeeded");
         Proposal storage proposal = proposals[proposalId];
-        uint eta = add256(block.timestamp, timelock.delay());
-        for (uint i = 0; i < proposal.targets.length; i++) {
+        uint256 eta = add256(block.timestamp, timelock.delay());
+        for (uint256 i = 0; i < proposal.targets.length; i++) {
             queueOrRevertInternal(proposal.targets[i], proposal.values[i], proposal.signatures[i], proposal.calldatas[i], eta);
         }
         proposal.eta = eta;
         emit ProposalQueued(proposalId, eta);
     }
 
-    function queueOrRevertInternal(address target, uint value, string memory signature, bytes memory data, uint eta) internal {
+    function queueOrRevertInternal(address target, uint256 value, string memory signature, bytes memory data, uint256 eta) internal {
         require(!timelock.queuedTransactions(keccak256(abi.encode(target, value, signature, data, eta))), "GovernorOHMega::queueOrRevertInternal: identical proposal action already queued at eta");
         timelock.queueTransaction(target, value, signature, data, eta);
     }
@@ -150,11 +150,11 @@ contract GovernorOHMegaDelegate is GovernorOHMegaDelegateStorageV1, GovernorOHMe
       * @notice Executes a queued proposal if eta has passed
       * @param proposalId The id of the proposal to execute
       */
-    function execute(uint proposalId) external payable {
+    function execute(uint256 proposalId) external payable {
         require(state(proposalId) == ProposalState.Queued, "GovernorOHMega::execute: proposal can only be executed if it is queued");
         Proposal storage proposal = proposals[proposalId];
         proposal.executed = true;
-        for (uint i = 0; i < proposal.targets.length; i++) {
+        for (uint256 i = 0; i < proposal.targets.length; i++) {
             timelock.executeTransaction.value(proposal.values[i])(proposal.targets[i], proposal.values[i], proposal.signatures[i], proposal.calldatas[i], proposal.eta);
         }
         emit ProposalExecuted(proposalId);
@@ -164,7 +164,7 @@ contract GovernorOHMegaDelegate is GovernorOHMegaDelegateStorageV1, GovernorOHMe
       * @notice Cancels a proposal only if sender is the proposer, or proposer delegates dropped below proposal threshold
       * @param proposalId The id of the proposal to cancel
       */
-    function cancel(uint proposalId) external {
+    function cancel(uint256 proposalId) external {
         require(state(proposalId) != ProposalState.Executed, "GovernorOHMega::cancel: cannot cancel executed proposal");
 
         Proposal storage proposal = proposals[proposalId];
@@ -172,7 +172,7 @@ contract GovernorOHMegaDelegate is GovernorOHMegaDelegateStorageV1, GovernorOHMe
         require(msg.sender == proposal.proposer || gOHM.getPriorVotes(proposal.proposer, sub256(block.number, 1)) < proposal.thresholdAtStart, "GovernorOHMega::cancel: proposer above threshold");
 
         proposal.canceled = true;
-        for (uint i = 0; i < proposal.targets.length; i++) {
+        for (uint256 i = 0; i < proposal.targets.length; i++) {
             timelock.cancelTransaction(proposal.targets[i], proposal.values[i], proposal.signatures[i], proposal.calldatas[i], proposal.eta);
         }
 
@@ -184,7 +184,7 @@ contract GovernorOHMegaDelegate is GovernorOHMegaDelegateStorageV1, GovernorOHMe
       * @param proposalId the id of the proposal
       * @return Targets, values, signatures, and calldatas of the proposal actions
       */
-    function getActions(uint proposalId) external view returns (address[] memory targets, uint[] memory values, string[] memory signatures, bytes[] memory calldatas) {
+    function getActions(uint256 proposalId) external view returns (address[] memory targets, uint256[] memory values, string[] memory signatures, bytes[] memory calldatas) {
         Proposal storage p = proposals[proposalId];
         return (p.targets, p.values, p.signatures, p.calldatas);
     }
@@ -195,7 +195,7 @@ contract GovernorOHMegaDelegate is GovernorOHMegaDelegateStorageV1, GovernorOHMe
       * @return votes
       */
       /// @notice change from original contract
-    function getVotesFromPercentOfsOHMSupply(uint percent) public view returns (uint256 votes) {
+    function getVotesFromPercentOfsOHMSupply(uint256 percent) public view returns (uint256 votes) {
         return gOHM.balanceTo(div256(mul256(sOHM.circulatingSupply(), percent), 1e6)); 
     }
 
@@ -205,7 +205,7 @@ contract GovernorOHMegaDelegate is GovernorOHMegaDelegateStorageV1, GovernorOHMe
       * @param voter The address of the voter
       * @return The voting receipt
       */
-    function getReceipt(uint proposalId, address voter) external view returns (Receipt memory) {
+    function getReceipt(uint256 proposalId, address voter) external view returns (Receipt memory) {
         return proposals[proposalId].receipts[voter];
     }
 
@@ -214,7 +214,7 @@ contract GovernorOHMegaDelegate is GovernorOHMegaDelegateStorageV1, GovernorOHMe
       * @param proposalId The id of the proposal
       * @return Proposal state
       */
-    function state(uint proposalId) public view returns (ProposalState) {
+    function state(uint256 proposalId) public view returns (ProposalState) {
         require(proposalCount >= proposalId && proposalId > initialProposalId, "GovernorOHMega::state: invalid proposal id");
         Proposal storage proposal = proposals[proposalId];
         if (proposal.canceled) {
@@ -242,7 +242,7 @@ contract GovernorOHMegaDelegate is GovernorOHMegaDelegateStorageV1, GovernorOHMe
       * @param proposalId The id of the proposal to vote on
       * @param support The support value for the vote. 0=against, 1=for, 2=abstain
       */
-    function castVote(uint proposalId, uint8 support) external {
+    function castVote(uint256 proposalId, uint8 support) external {
         emit VoteCast(msg.sender, proposalId, support, castVoteInternal(msg.sender, proposalId, support), "");
     }
 
@@ -252,7 +252,7 @@ contract GovernorOHMegaDelegate is GovernorOHMegaDelegateStorageV1, GovernorOHMe
       * @param support The support value for the vote. 0=against, 1=for, 2=abstain
       * @param reason The reason given for the vote by the voter
       */
-    function castVoteWithReason(uint proposalId, uint8 support, string calldata reason) external {
+    function castVoteWithReason(uint256 proposalId, uint8 support, string calldata reason) external {
         emit VoteCast(msg.sender, proposalId, support, castVoteInternal(msg.sender, proposalId, support), reason);
     }
 
@@ -260,7 +260,7 @@ contract GovernorOHMegaDelegate is GovernorOHMegaDelegateStorageV1, GovernorOHMe
       * @notice Cast a vote for a proposal by signature
       * @dev External function that accepts EIP-712 signatures for voting on proposals.
       */
-    function castVoteBySig(uint proposalId, uint8 support, uint8 v, bytes32 r, bytes32 s) external {
+    function castVoteBySig(uint256 proposalId, uint8 support, uint8 v, bytes32 r, bytes32 s) external {
         bytes32 domainSeparator = keccak256(abi.encode(DOMAIN_TYPEHASH, keccak256(bytes(name)), getChainIdInternal(), address(this)));
         bytes32 structHash = keccak256(abi.encode(BALLOT_TYPEHASH, proposalId, support));
         bytes32 digest = keccak256(abi.encodePacked("\x19\x01", domainSeparator, structHash));
@@ -276,14 +276,14 @@ contract GovernorOHMegaDelegate is GovernorOHMegaDelegateStorageV1, GovernorOHMe
       * @param support The support value for the vote. 0=against, 1=for, 2=abstain
       * @return The number of votes cast
       */
-    function castVoteInternal(address voter, uint proposalId, uint8 support) internal returns (uint) {
+    function castVoteInternal(address voter, uint256 proposalId, uint8 support) internal returns (uint256) {
         require(state(proposalId) == ProposalState.Active, "GovernorOHMega::castVoteInternal: voting is closed");
         require(support <= 2, "GovernorOHMega::castVoteInternal: invalid vote type");
         Proposal storage proposal = proposals[proposalId];
         Receipt storage receipt = proposal.receipts[voter];
         require(receipt.hasVoted == false, "GovernorOHMega::castVoteInternal: voter already voted");
         /// @notice change from original contract
-        uint votes = gOHM.getPriorVotes(voter, proposal.startBlock);
+        uint256 votes = gOHM.getPriorVotes(voter, proposal.startBlock);
 
         if (support == 0) {
             proposal.againstVotes = add256(proposal.againstVotes, votes);
@@ -304,10 +304,10 @@ contract GovernorOHMegaDelegate is GovernorOHMegaDelegateStorageV1, GovernorOHMe
       * @notice Admin function for setting the voting delay
       * @param newVotingDelay new voting delay, in blocks
       */
-    function _setVotingDelay(uint newVotingDelay) external {
+    function _setVotingDelay(uint256 newVotingDelay) external {
         require(msg.sender == admin, "GovernorOHMega::_setVotingDelay: admin only");
         require(newVotingDelay >= MIN_VOTING_DELAY && newVotingDelay <= MAX_VOTING_DELAY, "GovernorOHMega::_setVotingDelay: invalid voting delay");
-        uint oldVotingDelay = votingDelay;
+        uint256 oldVotingDelay = votingDelay;
         votingDelay = newVotingDelay;
 
         emit VotingDelaySet(oldVotingDelay,votingDelay);
@@ -317,10 +317,10 @@ contract GovernorOHMegaDelegate is GovernorOHMegaDelegateStorageV1, GovernorOHMe
       * @notice Admin function for setting the voting period
       * @param newVotingPeriod new voting period, in blocks
       */
-    function _setVotingPeriod(uint newVotingPeriod) external {
+    function _setVotingPeriod(uint256 newVotingPeriod) external {
         require(msg.sender == admin, "GovernorOHMega::_setVotingPeriod: admin only");
         require(newVotingPeriod >= MIN_VOTING_PERIOD && newVotingPeriod <= MAX_VOTING_PERIOD, "GovernorOHMega::_setVotingPeriod: invalid voting period");
-        uint oldVotingPeriod = votingPeriod;
+        uint256 oldVotingPeriod = votingPeriod;
         votingPeriod = newVotingPeriod;
 
         emit VotingPeriodSet(oldVotingPeriod, votingPeriod);
@@ -331,10 +331,10 @@ contract GovernorOHMegaDelegate is GovernorOHMegaDelegateStorageV1, GovernorOHMe
       * @dev newProposalThreshold must be greater than the hardcoded min
       * @param newProposalThreshold new proposal threshold
       */
-    function _setProposalThreshold(uint newProposalThreshold) external {
+    function _setProposalThreshold(uint256 newProposalThreshold) external {
         require(msg.sender == admin, "GovernorOHMega::_setProposalThreshold: admin only");
         require(newProposalThreshold >= MIN_PROPOSAL_THRESHOLDPERCENT && newProposalThreshold <= MAX_PROPOSAL_THRESHOLDPERCENT, "GovernorOHMega::_setProposalThreshold: invalid proposal threshold");
-        uint oldProposalThreshold = proposalThreshold;
+        uint256 oldProposalThreshold = proposalThreshold;
         proposalThreshold = newProposalThreshold;
 
         emit ProposalThresholdSet(oldProposalThreshold, proposalThreshold);
@@ -394,13 +394,13 @@ contract GovernorOHMegaDelegate is GovernorOHMegaDelegateStorageV1, GovernorOHMe
         emit NewPendingAdmin(oldPendingAdmin, pendingAdmin);
     }
 
-    function add256(uint256 a, uint256 b) internal pure returns (uint) {
-        uint c = a + b;
+    function add256(uint256 a, uint256 b) internal pure returns (uint256) {
+        uint256 c = a + b;
         require(c >= a, "addition overflow");
         return c;
     }
 
-    function sub256(uint256 a, uint256 b) internal pure returns (uint) {
+    function sub256(uint256 a, uint256 b) internal pure returns (uint256) {
         require(b <= a, "subtraction underflow");
         return a - b;
     }
@@ -424,8 +424,8 @@ contract GovernorOHMegaDelegate is GovernorOHMegaDelegateStorageV1, GovernorOHMe
         return c;
     }
 
-    function getChainIdInternal() internal pure returns (uint) {
-        uint chainId;
+    function getChainIdInternal() internal pure returns (uint256) {
+        uint256 chainId;
         assembly { chainId := chainid() }
         return chainId;
     }
