+++ contracts/interfaces/IOracle.sol
@@ -2,5 +2,5 @@
 pragma solidity 0.7.5;
 
 interface IOracle {
-    function getPrice( address _pool ) external returns ( uint );
+    function getPrice( address _pool ) external returns (uint256);
 }
